test: seting up tests with codecov
