+++ codecov.yml
@@ -1,0 +1,22 @@
+coverage:
+  status:
+    project: off
+    patch: off
+
+flag_management:
+  individual_flags:
+    - name: backend
+      paths:
+        - backend/
+      statuses:
+        - type: project
+          target: 80%
+          threshold: 1%
+
+    - name: frontend
+      paths:
+        - frontend/
+      statuses:
+        - type: project
+          target: 80%
+          threshold: 1%
