+++ frontend/__tests__/src/App.test.tsx
@@ -1,0 +1,6 @@
+import { render } from "@testing-library/react";
+import React from "react";
+import App from "../../src/App";
+test("renders the page", () => {
+  render(<App />);
+});
