+++ frontend/__tests__/src/components/fields/Pressure.test.tsx
@@ -1,19 +1,18 @@
-/* eslint-disable @typescript-eslint/no-empty-function */
-/* eslint-disable react/react-in-jsx-scope */
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import user from "@testing-library/user-event";
-import App from "../../App";
-describe("testing pathLength field ", () => {
-  test("testing pathLength field render with defaultValue and perfectly visible", () => {
+import React from "react";
+import App from "../../../../src/App";
+describe("testing pressure field ", () => {
+  test("testing pressure field render with defaultValue and perfectly visible", () => {
     render(<App />);
-    const input = screen.getByLabelText("Path Length");
+    const input = screen.getByLabelText("Pressure");
     expect(input).toBeVisible();
     expect(input).toBeInTheDocument();
-    expect(input).toHaveValue(1);
+    expect(input).toHaveValue(1.01325);
   });
-  test(" test pathLength field render with user* given value and perfectly visible", () => {
+  test(" test pressure field render with user* given value and perfectly visible", () => {
     render(<App />);
-    const input = screen.getByLabelText("Path Length");
+    const input = screen.getByLabelText("Pressure");
     expect(input).toBeVisible();
     expect(input).toBeInTheDocument();
     fireEvent.input(input, {
@@ -21,9 +20,9 @@ describe("testing pathLength field ", () => {
     });
     expect(input).toHaveValue(100);
   });
-  test("testing pathLength field validation for out of range values", async () => {
+  test("testing pressure field validation for out of range values", async () => {
     render(<App />);
-    const input = screen.getByLabelText("Path Length");
+    const input = screen.getByLabelText("Pressure");
     const button = screen.getByRole("button", {
       name: /new plot/i,
     });
@@ -33,22 +32,20 @@ describe("testing pathLength field ", () => {
     user.click(button);
     await waitFor(async () => {
       expect(
-        screen.getByText("Path length cannot be negative")
+        screen.getByText("Pressure cannot be negative")
       ).toBeInTheDocument();
     });
   });
   test("testing pressure field validation for undefined values", async () => {
     render(<App />);
-    const input = screen.getByLabelText("Path Length");
+    const input = screen.getByLabelText("Pressure");
     const button = screen.getByRole("button", {
       name: /new plot/i,
     });
     user.clear(input);
     user.click(button);
     await waitFor(async () => {
-      expect(
-        screen.getByText("Path length must be defined")
-      ).toBeInTheDocument();
+      expect(screen.getByText("Pressure must be defined")).toBeInTheDocument();
     });
   });
 });
