--- frontend/src/App.test.tsx
@@ -1,7 +1,0 @@
-/* eslint-disable @typescript-eslint/no-empty-function */
-/* eslint-disable react/react-in-jsx-scope */
-import { render } from "@testing-library/react";
-import App from "./App";
-test("renders the page", () => {
-  render(<App />);
-});
