+++ frontend/tsconfig.json
@@ -23,7 +23,9 @@
     "noUnusedParameters": true,
     "noImplicitReturns": true,
     "noImplicitAny": true,
+    "allowImportingTsExtensions" : true,
     "types": ["vitest/globals"]
+
   },
   "include": [
     "src"
