+++ frontend/vite.config.ts
@@ -1,5 +1,6 @@
 /// <reference types="vitest" />
 /// <reference types="vite/client" />
+
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 
@@ -19,7 +20,7 @@ export default defineConfig(() => {
       coverage: {
         provider: "v8",
         reporter: ["text", "json", "html"],
-        reportsDirectory: "./src/__tests__/coverage",
+        reportsDirectory: "./__tests__/coverage" /*coverage report*/,
       },
     },
     plugins: [react()],
